feat: update claim button
